Correct "choices" for DEVICE/MENU fields of enum record (#8)
When creating a pvAccess channel to a field of a record with a VAL field
of type DBF_ENUM whose DBF type is DBF_DEVICE or DBF_MENU, for example,
the DTYP or SCAN fields of a bi/bo/mbbi/mbbo record, and performing an
operation involving its "values.choices" field (Channel Get, Put or
Monitor) the "choices" field is filled in incorrectly with the values from the
record's "VAL" field. Fix this for EPICS base R3.14. Add equivalent
change for R3.15 which will also require fix for #17.

Fixes #8