Handle enum PVs when "value" not requested (#4)
Handle case when PV is an enum field and value is not asked for in
pvRequest.

Made slight change to meaning of bit mask: set enum, scalar and
scalarArray value bits according to type even if value not asked for.
Makes handling when "value" not asked for easier and more natural.
Allows fix in the case of monitors as well as gets. Requests for
timeStamps or alarms without values in gets, puts or monitors handled
correctly.

Fixes #4.