Fix error restarting monitors
When a monitor is started a new free monitorElement is obtained and
assigned to currentElement. However when this is a restart
currentElement is already holding onto a monitor element which will then
not be released. Stopping and restarting will exhaust the free monitor
elements and updates will stop.

Fix by not allocating a new element to currentElement each time, but
only when the latter is null.

Fixes #23.

(cherry picked from commit a35363b93831e4b46e316be1949c10cd5098cbf8)